Scale - updated weight from long to int

+++ src/KegMonitor.Core/Entities/Scale.cs
@@ -2,7 +2,7 @@
 {
     public class Scale : Entity
     {
-        public long CurrentWeight { get; private set; }
+        public int CurrentWeight { get; private set; }
         public int FullWeight { get; set; }
         public int EmptyWeight { get; set; }
 
@@ -14,7 +14,7 @@
         public IEnumerable<ScaleWeightChange> WeightChanges { get; private set; } = new List<ScaleWeightChange>();
         public DateTime LastUpdatedDated { get; set; }
 
-        public void UpdateWeight(long weight)
+        public void UpdateWeight(int weight)
         {
             CurrentWeight = weight;
             (WeightChanges as List<ScaleWeightChange>).Add(new ScaleWeightChange(this, weight));
